In some cases, symbolic information related to a struct/union might change. For example, we load DLKM-debug, or we need to use a different version of struct definition.
As normally this information is cached and persists until the end of crash
session, we need a way to invalidate it.

There is a new subroutine purge_typeinfo(sn = None). Calling it with
a specific struct name will purge caches for this struct only. Calling
this subroutine without arguments will purge all symbolic information.

As we cache even negative results, just checking whether struct exists
for non-existing struct will cache this. A typical scenario is deciding
whether we need to load an extra DLKM-debug. loadModule() now calls
purge_typeinfo() automatically.