rh_scsishow: Merge the get_gendev_* routines
Previously rh_scsishow.py was using separate routines
for getting a list of device's gendisk and request_queue
struct references. This patch merges it, so that we dont
need to do version checks for RHEL 5/6/7 kernels and just
call a single function which will return a list of
device's gendisk and corresponding request_queue struct
reference.