Add a check to verify Scsi_Host->hostt->module address
In some of the vmcores generated using most recent upstream
kernels, the Scsi_Host->hostt->module was found to be 0x0 when
the SCSI adapter was being managed by ata_piix driver.

In this case, trying to get the module version was resulting in
an error and call traces from print_shost_info routine in
scsishow.py

This patch adds a check to verify Scsi_Host->hostt->module address
before trying to further retrieve module version.