scsishow.py: Check for Qlogic HBA before retrieving fc_port
Before retrieving the fc_port structure add a check to verify if the
device is connected by Qlogic FC HBA. Adding this check earlier avoids
any attempt to retrieve fc_port struct for those devices which are not
connected by FC/FCoE HBAs. This would also remove a check for Qlogic
FC HBAs in later part of the code.