mountshow.py: handle NFS4 session related counts in --stats
Until recent kernels (4.19+, RHEL8) the NFS4 opcodes (SEQUENCE, RENEW, etc)
were not counted in /proc/self/mountstats.  This is now counted so
make sure we handle this here as well.

While we are converting this fix a bug with the handling of ktime_t
metrics by using the LinuxDump.Time library function ktime_t().

Signed-off-by: Dave Wysochanski <dwysocha@redhat.com>