Add new command kprobe_table.py to pykdump repository
 This commit adds a new command, kprobe_table.py, to the pykdump repository.
The command is designed to display all kprobes found in a Linux kernel dump.
It provides a convenient way to inspect and analyze kprobes within the
kernel dump, aiding in debugging and performance analysis.

 The kprobe_table.py command enhances the functionality of pykdump by
providing insights into the kprobes present in the kernel dump.
This information can be invaluable for understanding the behavior of
the kernel and diagnosing issues.

 This commit contributes to the ongoing development and enrichment of
the pykdump repository, empowering users with a comprehensive set of
tools for kernel dump analysis.

example usage:
```
epython pykdump/progs/kprobe_table.py
```

result:
```
Kprobe <struct kprobe 0xffff99668274a6c0>
  addr: <tPtr addr=0xffffffff858fc3a0 ctype='kprobe_opcode_t *'>
  opcode: 0
  pre_handler:
         addr 0xffffffff851ea4d0
         name aggr_pre_handler
  post_handler:
         addr 0x0
         name aggr_pre_handler
  symbol_name: None
  addr to sym: devlink_nl_param_fill
  offset: 0
  flags: 8
```

Signed-off-by: alexn1st0r <alexandernesterenko837@gmail.com>